+++ package.json
@@ -17,7 +17,9 @@
   },
   "homepage": "https://github.com/alvaropazz/js-tictactoe#readme",
   "devDependencies": {
-    "webpack": "^4.43.0",
+    "css-loader": "^4.2.1",
+    "style-loader": "^1.2.1",
+    "webpack": "^4.44.1",
     "webpack-cli": "^3.3.12"
   }
 }
